front-desk-telephony/pyproject.toml
[project]
name = "app1-front-desk-telephony"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=0.11.0",
    "livekit-plugins-google>=1.4.3",
    "livekit-plugins-openai>=1.4.3",
    "python-dotenv>=1.2.2",
]
